requirements.txt
PyGObject
ollama
duckduckgo-search
ddgs
beautifulsoup4
requests
lxml
openai
anthropic
google-genai
